requirements.txt
lxml>=4.9.3
psycopg2-binary>=2.9
